// src/types/line.ts
export declare type WebhookEvent =
  | MessageEvent
  | UnsendEvent
  | FollowEvent
  | UnfollowEvent
  | JoinEvent
  | LeaveEvent
  | MemberJoinEvent
  | MemberLeaveEvent
  | PostbackEvent
  | VideoPlayCompleteEvent
  | BeaconEvent
  | AccountLinkEvent
  | DeviceLinkEvent
  | DeviceUnlinkEvent
  | LINEThingsScenarioExecutionEvent;
export declare type EventBase = {
  /**
   * Channel state.
   *
   * `active`: The channel is active. You can send a reply message or push message from the bot server that received this webhook event.
   *
   * `standby`: The channel is waiting. The bot server that received this webhook event shouldn't send any messages.
   */
  mode: "active" | "standby";
  /**
   * Time of the event in milliseconds
   */
  timestamp: number;
  /**
   * Source user, group, or room object with information about the source of the event.
   */
  source: EventSource;
};
export declare type EventSource = User | Group | Room;
export declare type User = {
  type: "user";
  userId: string;
};
// ...
export declare type ReplyableEvent = EventBase & {
  replyToken: string;
};
/**
 * Webhook event object which contains the sent message.
 *
 * The `message` property contains a message object which corresponds with the
 * message type. You can reply to message events.
 *
 * @see [Message event](https://developers.line.biz/en/reference/messaging-api/#message-event)
 */
export declare type MessageEvent = {
  type: "message";
  message: EventMessage;
} & ReplyableEvent;
/**
 * Event object for when the user unsends a message in a [group](https://developers.line.biz/en/docs/messaging-api/group-chats/#group)
 * or [room](https://developers.line.biz/en/docs/messaging-api/group-chats/#room).
 * [Unsend event](https://developers.line.biz/en/reference/messaging-api/#unsend-event)
 */
export declare type UnsendEvent = {
  type: "unsend";
  /**
   * The message ID of the unsent message
   */
  unsend: {
    messageId: string;
  };
} & EventBase;
/**
 * Event object for when your account is added as a friend (or unblocked).
 */
export declare type FollowEvent = {
  type: "follow";
} & ReplyableEvent;
/**
 * Event object for when your account is blocked.
 */
export declare type UnfollowEvent = {
  type: "unfollow";
} & EventBase;
// ...
export declare type JoinEvent = {
  type: "join";
} & ReplyableEvent;
/**
 * Event object for when a user removes your bot from a group or a room.
 */
export declare type LeaveEvent = {
  type: "leave";
} & EventBase;
/**
 * Event object for when a user joins a [group](https://developers.line.biz/en/docs/messaging-api/group-chats/#group)
 * or [room](https://developers.line.biz/en/docs/messaging-api/group-chats/#room) that the bot is in.
 */
export declare type MemberJoinEvent = {
  type: "memberJoined";
  /**
   * User ID of users who joined
   * Array of [source user](https://developers.line.biz/en/reference/messaging-api/#source-user) objects
   */
  joined: {
    members: Array<User>;
  };
} & ReplyableEvent;
/**
 * Event object for when a user leaves a [group](https://developers.line.biz/en/docs/messaging-api/group-chats/#group)
 * or [room](https://developers.line.biz/en/docs/messaging-api/group-chats/#room) that the bot is in.
 */
export declare type MemberLeaveEvent = {
  type: "memberLeft";
  /**
   * User ID of users who left
   * Array of [source user](https://developers.line.biz/en/reference/messaging-api/#source-user) objects
   */
  left: {
    members: Array<User>;
  };
} & EventBase;
/**
 * Event object for when a user performs an action on a
 * [template message](https://developers.line.biz/en/reference/messaging-api/#template-messages).
 */
export declare type PostbackEvent = {
  type: "postback";
  postback: Postback;
} & ReplyableEvent;
/**
 * Event for when a user finishes viewing a video at least once with the specified trackingId sent by the LINE Official Account.
 */
export declare type VideoPlayCompleteEvent = {
  type: "videoPlayComplete";
  /**
   * ID used to identify a video. Returns the same value as the trackingId assigned to the [video message](https://developers.line.biz/en/reference/messaging-api/#video-message).
   * String
   */
  videoPlayComplete: {
    trackingId: string;
  };
} & ReplyableEvent;
/**
 * Event object for when a user enters or leaves the range of a
 * [LINE Beacon](https://developers.line.biz/en/docs/messaging-api/using-beacons/).
 */
export declare type BeaconEvent = ReplyableEvent & {
  type: "beacon";
  beacon: {
    /**
     * `leave` will be deprecated
     */
    type: "enter" | "leave" | "banner" | "stay";
    /**
     * Hardware ID of the beacon that was detected
     */
    hwid: string;
    /**
     * Device message of beacon that was detected.
     *
     * This message consists of data generated by the beacon to send notifications to bots.
     * Only included in webhooks from devices that support the "device message" property.
     * For more information, see the
     * [LINE Simple Beacon specification](https://github.com/line/line-simple-beacon/blob/master/README.en.md/#line-simple-beacon-frame).
     */
    dm?: string;
  };
};
/**
 * Event object for when a user has linked his/her LINE account with a provider's service account.
 */
export declare type AccountLinkEvent = ReplyableEvent & {
  type: "accountLink";
  link: {
    result: "ok" | "failed";
    /**
     * Specified nonce when verifying the user ID
     */
    nonce: string;
  };
};
/**
 * Indicates that a LINE Things-compatible device has been linked with LINE by a user operation.
 * For more information, see [Receiving device link events via webhook](https://developers.line.biz/en/docs/line-things/develop-bot/#link-event).
 */
export declare type DeviceLinkEvent = ReplyableEvent & {
  type: "things";
  things: {
    /**
     * Device ID of the LINE Things-compatible device that was linked with LINE
     */
    deviceId: string;
    type: "link";
  };
};
/**
 * Indicates that a LINE Things-compatible device has been unlinked from LINE by a user operation.
 * For more information, see [Receiving device unlink events via webhook](https://developers.line.biz/en/docs/line-things/develop-bot/#unlink-event).
 */
export declare type DeviceUnlinkEvent = ReplyableEvent & {
  type: "things";
  things: {
    /**
     * Device ID of the LINE Things-compatible device that was unlinked with LINE
     */
    deviceId: string;
    type: "unlink";
  };
};
export declare type LINEThingsScenarioExecutionEvent = ReplyableEvent & {
  type: "things";
  things: {
    type: "scenarioResult";
    /**
     * Device ID of the device that executed the scenario
     */
    deviceId: string;
    result: {
      /**
       * Scenario ID executed
       */
      scenarioId: string;
      /**
       * Revision number of the scenario set containing the executed scenario
       */
      revision: number;
      /**
       * Timestamp for when execution of scenario action started (milliseconds, LINE app time)
       */
      startTime: number;
      /**
       * Timestamp for when execution of scenario was completed (milliseconds, LINE app time)
       */
      endtime: number;
      /**
       * Scenario execution completion status
       * See also [things.resultCode definitions](https://developers.line.biz/en/reference/messaging-api/#things-resultcode).
       */
      resultCode: "success" | "gatt_error" | "runtime_error";
      /**
       * Execution result of individual operations specified in action
       * Note that an array of actions specified in a scenario has the following characteristics
       *  - The actions defined in a scenario are performed sequentially, from top to bottom.
       *  - Each action produces some result when executed.
       *    Even actions that do not generate data, such as `SLEEP`, return an execution result of type `void`.
       *    The number of items in an action array may be 0.
       *
       * Therefore, things.actionResults has the following properties:
       *  - The number of items in the array matches the number of actions defined in the scenario.
       *  - The order of execution results matches the order in which actions are performed.
       *    That is, in a scenario set with multiple `GATT_READ` actions,
       *    the results are returned in the order in which each individual `GATT_READ` action was performed.
       *  - If 0 actions are defined in the scenario, the number of items in things.actionResults will be 0.
       */
      actionResults: Array<LINEThingsActionResult>;
      /**
       * Data contained in notification
       * The value is Base64-encoded binary data.
       * Only included for scenarios where `trigger.type = BLE_NOTIFICATION`.
       */
      bleNotificationPayload?: string;
      /**
       * Error reason
       */
      errorReason?: string;
    };
  };
};
export declare type LINEThingsActionResult = {
  /**
   * `void`, `binary`
   * Depends on `type` of the executed action.
   * This property is always included if `things.actionResults` is not empty.
   */
  type: "void" | "binary";
  /**
   * Base64-encoded binary data
   *  This property is always included when `things.actionResults[].type` is `binary`.
   */
  data?: string;
};
export declare type EventMessage =
  | TextEventMessage
  | ImageEventMessage
  | VideoEventMessage
  | AudioEventMessage
  | LocationEventMessage
  | FileEventMessage
  | StickerEventMessage;
// ...
export declare type Postback = {
  data: string;
  params?: DateTimePostback | RichMenuSwitchPostback;
};